refactor: consolidate sidebar CSS and update to build b029
- Combined grid-layout-fix.css and sidebar-rail.css into sidebar.css
- Updated all asset version references to b029
- Simplified CSS loading order with z-overrides.css as last layout lock
- Clean separation of concerns: hero-cards, sidebar, then overrides
