+++ frontend/src/assets/css/sidebar.css
@@ -1,0 +1,107 @@
+/* ===== App grid shell ===== */
+:root{
+  --rail-cap: 280px;      /* max sidebar width */
+  --rail-gap: 0px;        /* column gap */
+}
+
+/* The two-column layout: rail auto-sizes to content (capped), main fills */
+.app{
+  display: grid !important;
+  grid-template-columns: fit-content(var(--rail-cap)) minmax(0,1fr) !important;
+  column-gap: var(--rail-gap) !important;
+  height: 100vh !important;
+}
+
+/* Sidebar: do NOT force a width; let the grid track size it */
+aside.sidebar,
+[data-role="sidebar"]{
+  position: sticky !important;
+  top: 0 !important;
+  height: 100vh !important;
+  transform: none !important;
+  /* remove any width/max-content lines you may have added before */
+}
+
+/* Main must scroll and be allowed to shrink */
+#main{
+  position: relative !important;
+  min-width: 0 !important;
+  height: 100vh !important;
+  overflow-y: auto !important;
+  overflow-x: hidden !important;
+  z-index: 0 !important;
+}:root{
+  --rail-accent-grad: linear-gradient(135deg,#8c7aff 0%,#5b47ff 55%,#3e7bff 100%);
+  --rail-ic-size: 28px;
+}
+
+/* Rail list column */
+.sidebar .nav{
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  padding: 12px 10px;
+}
+
+/* Single-line labels; never wrap; ellipsis for overflows within cap */
+.sidebar .nav a{
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  padding: 12px 14px;
+  padding-left: calc(var(--rail-ic-size) + 14px + 12px);
+  border: 1px solid var(--rail-border, #1f2536);
+  border-radius: 14px;
+  background: transparent;
+  color: #cfd8ff;
+  text-decoration: none;
+  position: relative;
+
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+
+  max-inline-size: var(--rail-cap);  /* belt-and-suspenders under the grid cap */
+}
+
+/* Virtual gradient chip icon */
+.sidebar .nav a::before{
+  content:"";
+  position:absolute;
+  left:14px; top:50%; transform:translateY(-50%);
+  width:var(--rail-ic-size); height:var(--rail-ic-size);
+  border-radius:12px;
+  background:var(--rail-accent-grad);
+  box-shadow: inset 0 0 0 1px rgba(255,255,255,.08), 0 6px 16px rgba(31,45,105,.35);
+  opacity:.85; transition:opacity .2s ease, transform .18s ease;
+}
+.sidebar .nav a:hover::before{ opacity:1; transform:translateY(-50%) scale(1.03); }
+.sidebar .nav a.is-active::before{ opacity:1; }
+
+/* Optional hash marker kept narrow and non-wrapping */
+.sidebar .nav a .hash{ opacity: .7; }
+
+/* If you later add <span class="ic">…</span> it overrides the ::before chip */
+.sidebar .nav a .ic{
+  width:var(--rail-ic-size); height:var(--rail-ic-size);
+  border-radius:12px; background:var(--rail-accent-grad);
+  box-shadow: inset 0 0 0 1px rgba(255,255,255,.08), 0 6px 16px rgba(31,45,105,.35);
+  display:inline-flex; align-items:center; justify-content:center;
+  margin-right:12px;
+}
+.sidebar .nav a .ic > svg{ width:60%; height:60%; fill:#fff; }
+
+/* Active accent (pairs with the gradient chip you liked) */
+.sidebar .nav a.is-active{
+  background: #111a33;
+  color: #eef3ff;
+  border-color: rgba(107,123,255,.35);
+  box-shadow: 0 0 0 2px rgba(107,123,255,.35), 0 14px 28px rgba(20,32,84,.45);
+}
+
+/* Hover focus */
+.sidebar .nav a:hover{ background: rgba(255,255,255,.03); }
+.sidebar .nav a:focus-visible{
+  outline: none;
+  box-shadow: 0 0 0 2px rgba(107,123,255,.35), 0 0 0 4px rgba(107,123,255,.25);
+}
