+++ frontend/src/js/router.js
@@ -1,5 +1,5 @@
-import { ensureShell, setActive } from './shell.js?v=b023';
-import "./gcal-hooks.js?v=b028";
+import { ensureShell, setActive } from './shell.js?v=b029';
+import "./gcal-hooks.js?v=b029";
 
 export const ROUTES = ['parties','calendar','map','hotspots','invites','contacts','me','settings'];
 export const currentRoute = () => (location.hash.replace(/^#\/?/, '')||'parties').split('?')[0];
@@ -12,15 +12,15 @@ async function go(){
   if (main) main.innerHTML = '';
 
   switch(r){
-    case 'parties':   (await import('./events-controller.js?v=b023')).renderParties(main); break;
-    case 'calendar':  (await import('./calendar-view.js?v=b023')).renderCalendar(main); break;
-    case 'contacts':  (await import('./contacts-panel.js?v=b023')).renderContacts?.(main); break;
-    case 'invites':   (await import('./invite-panel.js?v=b023')).renderInvites?.(main); break;
-    case 'map':       (await import('./map-controller.js?v=b023')).renderMap?.(main); break;
-    case 'hotspots':  (await import('./hotspots.js?v=b023')).renderHotspots?.(main); break;
-    case 'me':        (await import('./me-panel.js?v=b023')).renderMe?.(main); break;
-    case 'settings':  (await import('./settings-panel.js?v=b023')).renderSettings?.(main); break;
-    default:          (await import('./events-controller.js?v=b023')).renderParties(main);
+    case 'parties':   (await import('./events-controller.js?v=b029')).renderParties(main); break;
+    case 'calendar':  (await import('./calendar-view.js?v=b029')).renderCalendar(main); break;
+    case 'contacts':  (await import('./contacts-panel.js?v=b029')).renderContacts?.(main); break;
+    case 'invites':   (await import('./invite-panel.js?v=b029')).renderInvites?.(main); break;
+    case 'map':       (await import('./map-controller.js?v=b029')).renderMap?.(main); break;
+    case 'hotspots':  (await import('./hotspots.js?v=b029')).renderHotspots?.(main); break;
+    case 'me':        (await import('./me-panel.js?v=b029')).renderMe?.(main); break;
+    case 'settings':  (await import('./settings-panel.js?v=b029')).renderSettings?.(main); break;
+    default:          (await import('./events-controller.js?v=b029')).renderParties(main);
   }
 }
 window.addEventListener('hashchange', go);
